REFACTOR: NavBar component 분리
