+++ src/components/views/NavBar/NavBar.jsx
@@ -7,22 +7,11 @@ import AppBar from "@material-ui/core/AppBar";
 import Toolbar from "@material-ui/core/Toolbar";
 import IconButton from "@material-ui/core/IconButton";
 import Typography from "@material-ui/core/Typography";
-import MenuItem from "@material-ui/core/MenuItem";
-import Menu from "@material-ui/core/Menu";
-import MenuIcon from "@material-ui/icons/Menu";
 import AccountCircle from "@material-ui/icons/AccountCircle";
 import MoreIcon from "@material-ui/icons/MoreVert";
 
-import List from "@material-ui/core/List";
-import Divider from "@material-ui/core/Divider";
-import ListItem from "@material-ui/core/ListItem";
-import ListItemIcon from "@material-ui/core/ListItemIcon";
-import ListItemText from "@material-ui/core/ListItemText";
-import InboxIcon from "@material-ui/icons/MoveToInbox";
-import MailIcon from "@material-ui/icons/Mail";
-import Drawer from "@material-ui/core/Drawer";
-import { GoSearch } from "react-icons/go";
-import { BsCheckBox } from "react-icons/bs";
+import { DesktopMenu, MobileMenu, menuId, mobileMenuId } from "./Menu";
+import DrawerIcon from "./DrawerIcon";
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -70,8 +59,8 @@ const useStyles = makeStyles((theme) => ({
 
 const NavBar = () => {
   const classes = useStyles();
+
   const [anchorEl, setAnchorEl] = React.useState(null);
-  const [openDrawer, setOpenDrawer] = React.useState(false);
   const [mobileMoreAnchorEl, setMobileMoreAnchorEl] = React.useState(null);
 
   const isMenuOpen = Boolean(anchorEl);
@@ -94,105 +83,11 @@ const NavBar = () => {
     setMobileMoreAnchorEl(event.currentTarget);
   };
 
-  const toggleDrawer = (open) => (event) => {
-    if (
-      event.type === "keydown" &&
-      (event.key === "Tab" || event.key === "Shift")
-    ) {
-      return;
-    }
-
-    setOpenDrawer(open);
-  };
-
-  const menuId = "primary-search-account-menu";
-  const renderMenu = (
-    <Menu
-      anchorEl={anchorEl}
-      anchorOrigin={{ vertical: "top", horizontal: "right" }}
-      id={menuId}
-      keepMounted
-      transformOrigin={{ vertical: "top", horizontal: "right" }}
-      open={isMenuOpen}
-      onClose={handleMenuClose}
-    >
-      <Link to="/register" className={classes.sectionDesktopMenuItem}>
-        <MenuItem onClick={handleMenuClose}>회원가입</MenuItem>
-      </Link>
-      <Link to="/login" className={classes.sectionDesktopMenuItem}>
-        <MenuItem onClick={handleMenuClose}>로그인</MenuItem>
-      </Link>
-    </Menu>
-  );
-
-  const mobileMenuId = "primary-search-account-menu-mobile";
-  const renderMobileMenu = (
-    <Menu
-      anchorEl={mobileMoreAnchorEl}
-      anchorOrigin={{ vertical: "top", horizontal: "right" }}
-      id={mobileMenuId}
-      keepMounted
-      transformOrigin={{ vertical: "top", horizontal: "right" }}
-      open={isMobileMenuOpen}
-      onClose={handleMobileMenuClose}
-    >
-      <MenuItem onClick={handleProfileMenuOpen}>
-        <IconButton
-          aria-label="account of current user"
-          aria-controls="primary-search-account-menu"
-          aria-haspopup="true"
-          color="inherit"
-        >
-          <AccountCircle />
-        </IconButton>
-        <p>Profile</p>
-      </MenuItem>
-    </Menu>
-  );
-
-  const list = (
-    <div
-      className={classes.list}
-      role="presentation"
-      onClick={toggleDrawer(false)}
-      onKeyDown={toggleDrawer(false)}
-    >
-      <List>
-        <ListItem button>
-          <ListItemIcon>
-            <GoSearch />
-          </ListItemIcon>
-          <ListItemText primary="유치원 검색" />
-        </ListItem>
-
-        <ListItem button>
-          <ListItemIcon>
-            <BsCheckBox />
-          </ListItemIcon>
-          <ListItemText primary="유치원 리뷰 조회" />
-        </ListItem>
-      </List>
-
-      <Divider />
-    </div>
-  );
-
   return (
     <div className={classes.grow}>
       <AppBar position="static">
         <Toolbar className="container">
-          <IconButton
-            edge="start"
-            className={classes.menuButton}
-            color="inherit"
-            aria-label="open drawer"
-            onClick={toggleDrawer(true)}
-          >
-            <MenuIcon />
-          </IconButton>
-          <Drawer anchor="left" open={openDrawer} onClose={toggleDrawer(false)}>
-            {list}
-          </Drawer>
+          <DrawerIcon classes={classes} />
 
           <Typography className={classes.title} variant="h6" noWrap>
             <Link to="/" className={classes.title}>
@@ -226,8 +121,19 @@ const NavBar = () => {
           </div>
         </Toolbar>
       </AppBar>
-      {renderMobileMenu}
-      {renderMenu}
+
+      <DesktopMenu
+        classes={classes}
+        anchorEl={anchorEl}
+        isMenuOpen={isMenuOpen}
+        handleMenuClose={handleMenuClose}
+      />
+      <MobileMenu
+        mobileMoreAnchorEl={mobileMoreAnchorEl}
+        isMobileMenuOpen={isMobileMenuOpen}
+        handleMobileMenuClose={handleMobileMenuClose}
+        handleProfileMenuOpen={handleProfileMenuOpen}
+      />
     </div>
   );
 };
